chore: no more speculate tests
This patch ports the `speculate` tests to standard rust tests. I
specifically did not try to address #174 because this patch is big
enough as it is, without any real functional changes. This is a
relatively direct port, so doesn't have any functionality changes in it.

Fixes #215
