+++ Cargo.toml
@@ -34,6 +34,3 @@ url = "2.2.2"
 wasm-bindgen = "0.2.74"
 wasm-bindgen-futures = "0.4.24"
 web-sys = { version = "0.3.51", features = ["console"] }
-
-[dev-dependencies]
-speculate = "0.1.2"
